E1_gestion_donnees/requirements.txt
scrapy
scrapy-playwright
pandas
numpy
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiohttp
Pillow
python-dotenv
